Merge release-v1.4.100 into master for v1.4.100 release

+++ scripts/build.sh
@@ -4,9 +4,17 @@
 SDK_VERSION_STR=`grep "^SDK_VERSION=" ${SDK_PATH}/sdk/tools/mkversion.sh | cut -d "\"" -f 2 | sed s/SDK//g`
 SDK_VERSION_MAJ=`echo ${SDK_VERSION_STR} | cut -d "." -f 1`
 
-# NuttX Version check
-NTX_VERSION_STR=`grep "^NUTTX_VERSION=" ${SDK_PATH}/sdk/tools/mkversion.sh | cut -d "\"" -f 2`
-NTX_VERSION_MAJ=`echo ${NTX_VERSION_STR} | cut -d "." -f 1`
+# Import NuttX version information
+#
+# It must be CONFIG_VERSION_{STRING|MAJOR|MINOR|PATCH|BUILD}.
+# Do not redefine these variables in this script.
+if [ -e ${SDK_PATH}/nuttx/.version ]; then
+    . ${SDK_PATH}/nuttx/.version
+else
+    CONFIG_VERSION_MAJOR=0
+    CONFIG_VERSION_MINOR=0
+    CONFIG_VERSION_PATCH=0
+fi
 
 # Location of .config
 LOCAL_NUTTX_CONFIG=${SDK_PATH}/nuttx/.config
@@ -93,9 +101,11 @@ function build_sdk (){
     if [ "${SDK_VERSION_MAJ}" == "1" ]; then
         rm -f ${SDK_PATH}/sdk/system/builtin/registry/.updated
     else
-        rm -f ${SDK_PATH}/sdk/apps/builtin/registry/.updated
-        rm -f ${SDK_PATH}/sdk/apps/builtin/registry/*.bdat
-        rm -f ${SDK_PATH}/sdk/apps/builtin/registry/*.pdat
+        if [ $CONFIG_VERSION_MAJOR -le 10 ]; then
+            rm -f ${SDK_PATH}/sdk/apps/builtin/registry/.updated
+            rm -f ${SDK_PATH}/sdk/apps/builtin/registry/*.bdat
+            rm -f ${SDK_PATH}/sdk/apps/builtin/registry/*.pdat
+        fi
     fi
 
     # Check SDK config
